ref(sdk-crashes): Rename Flutter files to Dart
Rename Flutter files to Dart, as all other files and methods are named with Dart.
